Refactor hot-reload: move to cli module, remove pyproject change, add dag-processor support
