+++ airflow-core/pyproject.toml
@@ -140,7 +140,6 @@ dependencies = [
     # https://github.com/apache/airflow/issues/56369 , rework universal-pathlib usage
     "universal-pathlib>=0.2.6,<0.3.0",
     "uuid6>=2024.7.10",
-    "watchfiles>=0.20.0",
     "apache-airflow-task-sdk<1.3.0,>=1.1.0",
     # pre-installed providers
     "apache-airflow-providers-common-compat>=1.7.4",
