+++ airflow-core/src/airflow/cli/commands/dag_processor_command.py
@@ -52,6 +52,13 @@ def dag_processor(args):
     """Start Airflow Dag Processor Job."""
     job_runner = _create_dag_processor_job_runner(args)
 
+    if hasattr(args, "dev") and args.dev:
+        log.info("Starting dag-processor in development mode with hot-reload enabled")
+        from airflow.cli.hot_reload import run_with_reloader
+
+        run_with_reloader(lambda: run_job(job=job_runner.job, execute_callable=job_runner._execute))
+        return
+
     run_command_with_daemon_option(
         args=args,
         process_name="dag-processor",
